utils for main page implemented
